+++ CHANGELOG.md
@@ -43,6 +43,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Animated progress bar during initialization and connecting
 - New manifest format support
 - Sending 'patcher_started' event to Statistics Reporting Service
+- Custom building options under `Tools/Build`
 - Handling failures when accessing the file system
 - Handling torrent-client crashes
 - Attaching "system-info" to Sentry events as tag
@@ -62,6 +63,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Availability of user action buttons (update, start & check for updates)
 - Handling of the ZLib exception
 - Invalid handling of patcher-data-location argument with spaces
+
+### Removed
+- StandaloneOSXUniversal architecture from building options
 
 ## [3.9.2]
 ### Added